Renaming files and fixing Sidenav selected prop when new section is clicked
