+++ website/pages/utilities/halstack-provider.tsx
@@ -1,5 +1,5 @@
 import Head from "next/head";
-import HalstackProviderPage from "../../screens/utilities/halstackProvider/HalstackProviderPage";
+import HalstackProviderPage from "../../screens/utilities/halstack-provider/HalstackProviderPage";
 
 const Index = () => {
   return (
